backend/requirements.txt
channels-redis==4.2.0
daphne==4.1.2
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.15.1
fqdn==1.5.1
isoduration==20.11.0
jsonpointer==2.4
notebook==6.5.6
pip-chill==1.0.3
pyopenssl==24.1.0
service-identity==24.1.0
uri-template==1.3.0
webcolors==1.13
zope.interface==6.4
